% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwc_colors.R
\name{pwc_colors}
\alias{pwc_colors}
\title{PwC Colors}
\usage{
pwc_colors()
}
\value{
Character vector of color hex codes
}
\description{
Retrieve a vector specifying the custom PwC color palette generated off the
official logo using \code{\link[paletter:create_palette]{paletter::create_palette()}}.
}
\examples{
library(pwctools)
pwc_colors()
}
\keyword{colors}
