% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwctools-package.R
\docType{package}
\name{pwctools-package}
\alias{pwctools}
\alias{pwctools-package}
\title{pwctools: PwC Toolkit}
\description{
Internal tools for PwC Developers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jimbrig/pwctools}
  \item \url{https://jimbrig.github.io/pwctools/}
  \item Report bugs at \url{https://github.com/jimbrig/pwctools/issues}
}

}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
