% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwc_colors.R
\name{view_pwc_colors}
\alias{view_pwc_colors}
\title{View PwC Colors}
\usage{
view_pwc_colors(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[scales:show_col]{scales::show_col}}
  \describe{
    \item{\code{colours}}{A character vector of colours}
    \item{\code{labels}}{Label each colour with its hex name?}
    \item{\code{borders}}{Border colour for each tile. Default uses \code{par("fg")}.
Use \code{border = NA} to omit borders.}
    \item{\code{cex_label}}{Size of printed labels, as multiplier of default size.}
    \item{\code{ncol}}{Number of columns. If not supplied, tries to be as square as
possible.}
  }}
}
\description{
View the PwC Color Palette.
}
\examples{
\dontrun{
view_pwc_colors()
}
}
\keyword{colors}
